requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.8.30
cffi==1.17.1
cryptography==43.0.1
dnspython==2.6.1
pycparser==2.22
pymongo==4.9.1
python-telegram-bot==13.14
pytz==2024.2
six==1.16.0
tornado==6.1
tzlocal==5.2
